+++ _script/package.sh
@@ -250,7 +250,6 @@ TEXTURES_NOT_FOR_VANILLA=(
 master_marksman
 molten_core
 art_of_war
-sudden_death
 shooting_stars
 daybreak
 backlash